[android] Moved all map tasks to intent package
