+++ android/src/com/mapswithme/maps/intent/MapTask.java
@@ -1,0 +1,12 @@
+package com.mapswithme.maps.intent;
+
+import android.support.annotation.NonNull;
+
+import com.mapswithme.maps.MwmActivity;
+
+import java.io.Serializable;
+
+public interface MapTask extends Serializable
+{
+  boolean run(@NonNull MwmActivity target);
+}
